subgraphs/members/requirements.txt
cachetools==6.2.1
fastapi==0.116.1
httpx==0.28.1
pymongo==4.14.1
strawberry-graphql[debug-server]==0.281.0
uvicorn==0.35.0
